float_basic.c:    - FloatsEqual(): Use fabsf to avoid double. arith_x86_gcc/arith_x86_gcc.cbp:    - Adds path to $SPJ_SWR_LOC)/libs_common/public/spj_stdint.h arith_ccs_msp430/.cproject:    - Adds --fp_mode=relaxed for more efficient FP.

+++ src/float_basic.c
@@ -301,9 +301,9 @@ PUBLIC BOOL FloatsEqual(float a, float b, float epsilon) {
       return FALSE; }
    else
    {
-      float absA = fabs(a);
-      float absB = fabs(b);
-      float diff = fabs(a - b);
+      float absA = fabsf(a);
+      float absB = fabsf(b);
+      float diff = fabsf(a - b);
 
       if (a == b) {                                               // shortcut, handles infinities
          return TRUE; }
